% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaPlot.R
\name{addCpaMarkersToPlot}
\alias{addCpaMarkersToPlot}
\title{adds CPA markers to plot,
use read-tv with CPA, then click "View Source" on the CPA tab AFTER running CPA to see an example.}
\usage{
addCpaMarkersToPlot(
  time_plot,
  cpa_df,
  plot_data,
  time_column = "Time",
  y_column = NULL,
  facet_column = NULL
)
}
\description{
adds CPA markers to plot,
use read-tv with CPA, then click "View Source" on the CPA tab AFTER running CPA to see an example.
}
