% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaUtils.R
\name{cpaResultsToDataFrame}
\alias{cpaResultsToDataFrame}
\title{This simple function takes the output of a changepoint algorithm,
with a time array, and creates a data frame of 2 columns:
\enumerate{
\item cpts (change points, in time, the final time point before the change)
\item vals (the calculated average \alue of the data before the change point)
}}
\usage{
cpaResultsToDataFrame(cpt_res, time_data)
}
\description{
This simple function takes the output of a changepoint algorithm,
with a time array, and creates a data frame of 2 columns:
\enumerate{
\item cpts (change points, in time, the final time point before the change)
\item vals (the calculated average \alue of the data before the change point)
}
}
