% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customizeDisplayModule.R
\name{generatePlotDefaults}
\alias{generatePlotDefaults}
\title{Generate Plot Defaults for Basic Time Plot}
\usage{
generatePlotDefaults(overrides = list())
}
\description{
This function is for internal use. It is exported so that users can view key-value pairs used for plotting. Reveals key names and value defaults. See \code{help(tvOpts)}.
}
\details{
See \code{help(tvOpts)} for more information. Should not be used for other reasons besides viewing plot defaults and param names.
}
\examples{
generatePlotDefaults()
}
