% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchReadtv}
\alias{launchReadtv}
\title{Launch or create READ-TV shiny application}
\usage{
launchReadtv(data = NULL, eventsPath = NULL, plotOpts = list())
}
\arguments{
\item{data}{A data frame. If \code{data} does not have columns for \code{Time}, \code{Case}, and \code{Event.Type}, then a pop-up will display to map these columns.}

\item{eventsPath}{A file path to a data frame. Can not be set with \code{data}. READ-TV will load \code{eventsPath} with \code{readr::read_rds} or \code{readr::read_csv}.}

\item{plotOpts}{A list of keys and values for generating the first basic display plot. THe defaults can be viewed with \code{generatePlotDefaults()}. Wrap in \code{tvOpts} call for non-standard evaluation. See \code{?tvOpts} for more information.}
}
\description{
Returns a read-tv shiny application, or launches it if not set
}
\examples{

library(dplyr)
library(magrittr)

app = readtv::japan_eq_3_11 \%>\%
  mutate(Time = time, Case = 1, Event.Type = place) \%>\%
  launchReadtv(plotOpts = tvOpts(
    x = time, y = mag, color = place, facetOn = place,
    isFacetPaginated = TRUE, facetRowsPerPage = 3, facetPage = 2))

#shiny::runApp(app, port = 80, host = '0.0.0.0')

}
