% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tvOpts}
\alias{tvOpts}
\title{Use Non-standard evaluation with functions that require a list with string, integer, double, or bool values}
\usage{
tvOpts(...)
}
\value{
list with string, integer, double, or bool values
}
\description{
This is used with \code{launchReadtv} (\code{plotOpts} argument).
\code{launchReadtv} expects a list with non-expression values.
This type of list can be generated with \code{tvOpts}
}
\details{
For any values not set, \code{launchReadtv} will use the defaults found in \code{generatePlotDefaults()}.
\code{generatePlotDefaults()} will also show which list names can be set, other names will be ignored.
}
\examples{

library(dplyr)
library(magrittr)

readtv::japan_eq_3_11 \%>\%
  with(tvOpts(x = time, y = mag, facetOn = place))

app = readtv::japan_eq_3_11 \%>\%
  mutate(Time = time, Case = 1, Event.Type = place) \%>\%
  launchReadtv(plotOpts = tvOpts(
    x = time, y = mag, color = place, facetOn = place,
    isFacetPaginated = TRUE, facetRowsPerPage = 3, facetPage = 2))

#shiny::runApp(app)

}
